added services

+++ modules/admin/http/controllers/services/serviceController.js
@@ -8,6 +8,7 @@ const {
   sendSuccessResponse,
   sendErrorResponse,
   sendNotFoundError,
+  sendValidationError,
 } = require("../../traits/responseHandler");
 const { paginate } = require("../../traits/datatablePaginationHelper");
 const {
@@ -19,62 +20,66 @@ const {
 const {
   validationRequestPost,
   validateId,
-} = require("../../request/contact/connectionsRequest");
-const {
-  updateValidationCheck,
-  createValidationCheck,
-  idValidationCheck,
-} = require("../../traits/validationHelper");
+} = require("../../request/services/serviceRequest");
+const { validationResult } = require("express-validator");
 
-const dataModel = models.Connections;
-const fileFields = ["icon_media_path"];
+const dataModel = models.Services;
 
-class ConnectionsController {
+class ServiceController {
   static async list(req, res) {
     try {
-      const listCacheKey = cacheKeys.connectionsList(req);
+      const listCacheKey = cacheKeys.serviceList(req);
       const cached = await getCache(req, listCacheKey);
       if (cached) {
-        return sendSuccessResponse(res, cached, "Connections retrieved successfully from cache");
+        return sendSuccessResponse(res, cached, "Services retrieved successfully from cache");
       }
+
       const result = await paginate(dataModel, req, {
         order: [["sort_order", "ASC"]],
-        searchFields: ["title", "description"],
+        searchFields: ["title"], 
       });
+
       await setCache(req, listCacheKey, result);
-      sendSuccessResponse(res, result, "Connections retrieved successfully");
+      sendSuccessResponse(res, result, "Services retrieved successfully");
     } catch (error) {
       return sendErrorResponse(res, error);
     }
   }
 
   static async getById(req, res) {
-    await idValidationCheck(req, res, validateId);
+    await Promise.all(validateId.map((validation) => validation.run(req)));
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) return sendValidationError(res, errors.array());
+
     try {
       const { id } = req.params;
-      const itemCacheKey = cacheKeys.connectionsItem(id);
+      const itemCacheKey = cacheKeys.serviceItem(id);
       const cached = await getCache(req, itemCacheKey);
-      if (cached) return sendSuccessResponse(res, cached, "Connection retrieved successfully");
+      if (cached) return sendSuccessResponse(res, cached, "Service retrieved successfully");
 
       const item = await dataModel.findByPk(id);
-      if (!item) return sendNotFoundError(res, "Connection");
+      if (!item) return sendNotFoundError(res, "Service");
 
       await setCache(req, itemCacheKey, item);
-      sendSuccessResponse(res, item, "Connection retrieved successfully");
+      sendSuccessResponse(res, item, "Service retrieved successfully");
     } catch (error) {
       return sendErrorResponse(res, error);
     }
   }
 
   static async create(req, res) {
-    await createValidationCheck(req, res, validationRequestPost);
+    await Promise.all(validationRequestPost.map((v) => v.run(req)));
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) return sendValidationError(res, errors);
+
     const t = await sequelize.transaction();
     try {
-      handleFileUploadStore(req, fileFields);
       const item = await dataModel.create(req.body, { transaction: t });
+
       await t.commit();
-      await invalidateCache(req, cacheKeys.connectionsListPattern());
-      sendSuccessResponse(res, item, "Connection created successfully", 201);
+
+      await invalidateCache(req, cacheKeys.serviceListPattern());
+      sendSuccessResponse(res, item, "Service created successfully", 201);
     } catch (error) {
       await t.rollback();
       return sendErrorResponse(res, error);
@@ -82,21 +87,27 @@ class ConnectionsController {
   }
 
   static async update(req, res) {
-    await updateValidationCheck(req, res, validationRequestPost, validateId);
+    await Promise.all([...validateId, ...validationRequestPost].map((v) => v.run(req)));
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) return sendValidationError(res, errors.array());
+
     const t = await sequelize.transaction();
     try {
       const { id } = req.params;
       const item = await dataModel.findByPk(id);
       if (!item) {
         await t.rollback();
-        return sendNotFoundError(res, "Connection");
+        return sendNotFoundError(res, "Service");
       }
-      await handleFileUploadUpdate(req, item, fileFields);
+
       await item.update(req.body, { transaction: t });
+
       await t.commit();
-      await invalidateCache(req, cacheKeys.connectionsItem(id));
-      await invalidateCache(req, cacheKeys.connectionsListPattern());
-      sendSuccessResponse(res, item, "Connection updated successfully");
+
+      await invalidateCache(req, cacheKeys.serviceItem(id));
+      await invalidateCache(req, cacheKeys.serviceListPattern());
+
+      sendSuccessResponse(res, item, "Service updated successfully");
     } catch (error) {
       await t.rollback();
       return sendErrorResponse(res, error);
@@ -104,21 +115,29 @@ class ConnectionsController {
   }
 
   static async destroy(req, res) {
-    await idValidationCheck(req, res, validateId);
+    await Promise.all(validateId.map((validation) => validation.run(req)));
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) return sendValidationError(res, errors.array());
+
     const t = await sequelize.transaction();
     try {
       const { id } = req.params;
       const item = await dataModel.findByPk(id);
       if (!item) {
         await t.rollback();
-        return sendNotFoundError(res, "Connection");
+        return sendNotFoundError(res, "Service");
       }
+
       await item.destroy({ transaction: t });
+
       await t.commit();
-      if (item.icon_media_path) await deleteOldFile(item.icon_media_path);
-      await invalidateCache(req, cacheKeys.connectionsItem(id));
-      await invalidateCache(req, cacheKeys.connectionsListPattern());
-      sendSuccessResponse(res, { id: id }, "Connection deleted successfully");
+
+      if (item.profile_media_path) await deleteOldFile(item.profile_media_path);
+
+      await invalidateCache(req, cacheKeys.serviceItem(id));
+      await invalidateCache(req, cacheKeys.serviceListPattern());
+
+      sendSuccessResponse(res, { id: id }, "Service deleted successfully");
     } catch (error) {
       await t.rollback();
       return sendErrorResponse(res, error);
@@ -126,4 +145,4 @@ class ConnectionsController {
   }
 }
 
-module.exports = ConnectionsController;
+module.exports = ServiceController;
